fix: SonarCloud 경고 2차 수정 - 공통 components 생성
