+++ src/features/place/screens/EditPlaceScreen.tsx
@@ -16,6 +16,8 @@ import React, { useState } from 'react'
 import { Shadow } from 'react-native-shadow-2'
 import { Ionicons } from '@expo/vector-icons'
 import DateTimePicker from '@react-native-community/datetimepicker'
+import NoiseOverlay from '@/src/components/common/NoiseOverlay'
+import Dropdown from '@/src/components/common/Dropdown'
 
 const { width } = Dimensions.get('window')
 const CARD_WIDTH = width * 0.91
@@ -29,19 +31,6 @@ const REGIONS = [
     '용산구', '은평구', '종로구', '중구', '중랑구'
 ]
 
-const NoiseOverlay = () => (
-    <>
-        <Image
-            source={require('../../../../assets/images/noise.png')}
-            style={[styles.noiseOverlay, { top: 0 }]}
-        />
-        <Image
-            source={require('../../../../assets/images/noise.png')}
-            style={[styles.noiseOverlay, { top: 400 }]}
-        />
-    </>
-)
-
 const InfoRow = ({ iconName, text }: { iconName: string, text: string }) => (
     <View style={styles.infoRowWrapper}>
         <View style={styles.infoRow}>
@@ -51,72 +40,6 @@ const InfoRow = ({ iconName, text }: { iconName: string, text: string }) => (
         <View style={styles.divider} />
     </View>
 )
-
-const Dropdown = ({
-    label,
-    value,
-    options,
-    onSelect
-}: {
-    label: string
-    value: string
-    options: string[]
-    onSelect: (v: string) => void
-}) => {
-    const [open, setOpen] = useState(false)
-
-    return (
-        <View style={styles.dropdownWrapper}>
-            <Text style={styles.dropdownLabel}>{label}</Text>
-            <TouchableOpacity style={styles.dropdown} onPress={() => setOpen(true)}>
-                <Text style={styles.dropdownText}>{value}</Text>
-                <Ionicons name={open ? 'chevron-up' : 'chevron-down'} size={16} color="#333" />
-            </TouchableOpacity>
-
-            <Modal
-                visible={open}
-                transparent
-                animationType="fade"
-                onRequestClose={() => setOpen(false)}
-            >
-                <TouchableOpacity
-                    style={styles.modalBackdrop}
-                    onPress={() => setOpen(false)}
-                    activeOpacity={1}
-                >
-                    <View style={styles.modalBox}>
-                        <FlatList
-                            data={options}
-                            keyExtractor={(item) => item}
-                            renderItem={({ item }) => (
-                                <TouchableOpacity
-                                    style={[
-                                        styles.modalItem,
-                                        item === value && styles.modalItemSelected
-                                    ]}
-                                    onPress={() => {
-                                        onSelect(item)
-                                        setOpen(false)
-                                    }}
-                                >
-                                    <Text style={[
-                                        styles.modalItemText,
-                                        item === value && styles.modalItemTextSelected
-                                    ]}>
-                                        {item}
-                                    </Text>
-                                    {item === value && (
-                                        <Ionicons name="checkmark" size={16} color="#1A3A6B" />
-                                    )}
-                                </TouchableOpacity>
-                            )}
-                        />
-                    </View>
-                </TouchableOpacity>
-            </Modal>
-        </View>
-    )
-}
 
 type Props = {
     onBack: () => void
@@ -351,15 +274,6 @@ const styles = StyleSheet.create({
         marginTop: 20, 
     },
 
-    noiseOverlay: {
-        position: 'absolute',
-        top: 0,
-        left: 0,
-        right: 0,
-        bottom: 0,
-        opacity: 1,
-    },   
-
     logContainer: {
         height: 705,
         borderRadius: 16,
@@ -435,74 +349,11 @@ const styles = StyleSheet.create({
         marginTop: 16,
     },
 
-    dropdownWrapper: {
-        flex: 1,
-        gap: 6,
-    },
-
-    dropdownLabel: {
-        fontSize: 13,
-        color: '#555',
-        fontWeight: '500',
-    },
-
-    dropdown: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        borderWidth: 1.5,
-        borderColor: '#1A3A6B',
-        backgroundColor: '#ffffff',
-        borderRadius: 12,
-        paddingHorizontal: 14,
-        paddingVertical: 10,
-    },
-
-    dropdownText: {
-        fontSize: 14,
-        color: '#222',
-    },
-
     modalBackdrop: {
         flex: 1,
         backgroundColor: '#00000040',
         justifyContent: 'center',
         alignItems: 'center',
-    },
-
-    modalBox: {
-        width: 200,
-        maxHeight: 320,
-        backgroundColor: '#fff',
-        borderRadius: 16,
-        paddingVertical: 8,
-        elevation: 10,
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 4 },
-        shadowOpacity: 0.15,
-        shadowRadius: 10,
-    },
-
-    modalItem: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        paddingHorizontal: 16,
-        paddingVertical: 12,
-    },
-
-    modalItemSelected: {
-        backgroundColor: '#F0F4FF',
-    },
-
-    modalItemText: {
-        fontSize: 14,
-        color: '#333',
-    },
-
-    modalItemTextSelected: {
-        color: '#1A3A6B',
-        fontWeight: '600',
     },
 
     contentSection: {
